backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
motor==3.5.1
pydantic==2.8.2
pydantic-settings==2.4.0
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
certifi==2024.8.30
email-validator==2.2.0
